// apps/web/src/api/generated/tournamentEligibilityIssueDtoCode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type TournamentEligibilityIssueDtoCode = typeof TournamentEligibilityIssueDtoCode[keyof typeof TournamentEligibilityIssueDtoCode];


export const TournamentEligibilityIssueDtoCode = {
  CAPTAIN_PROFILE_INCOMPLETE: 'CAPTAIN_PROFILE_INCOMPLETE',
  TEAM_NOT_CREATED: 'TEAM_NOT_CREATED',
  TEAM_INACTIVE: 'TEAM_INACTIVE',
  REGISTRATION_NOT_OPEN: 'REGISTRATION_NOT_OPEN',
  REGISTRATION_DEADLINE_PASSED: 'REGISTRATION_DEADLINE_PASSED',
  TOURNAMENT_FULL: 'TOURNAMENT_FULL',
  ALREADY_REGISTERED: 'ALREADY_REGISTERED',
  GAME_MISMATCH: 'GAME_MISMATCH',
  INSUFFICIENT_STARTERS: 'INSUFFICIENT_STARTERS',
  TOO_MANY_STARTERS: 'TOO_MANY_STARTERS',
  TOO_MANY_SUBSTITUTES: 'TOO_MANY_SUBSTITUTES',
  MISSING_GAME_ACCOUNT_ID: 'MISSING_GAME_ACCOUNT_ID',
  MISSING_PLAYER_PHONE: 'MISSING_PLAYER_PHONE',
  REGION_NOT_ALLOWED: 'REGION_NOT_ALLOWED',
  COUNTRY_NOT_ALLOWED: 'COUNTRY_NOT_ALLOWED',
  PLATFORM_NOT_ALLOWED: 'PLATFORM_NOT_ALLOWED',
  RANK_NOT_ALLOWED: 'RANK_NOT_ALLOWED',
  PLAYER_INELIGIBLE: 'PLAYER_INELIGIBLE',
} as const;